+++ src/sdsstools/__init__.py
@@ -13,12 +13,8 @@ NAME = 'sdsstools'
 from .configuration import *
 from .logger import *
 from .metadata import *
-from ._vendor import color_print, releases, toml
+from ._vendor import color_print, toml
 
 # This is a hack to allow doing from sdsstools.color_print import color_text
 # which some code already does.
 sys.modules['sdsstools.color_print'] = color_print
-
-# Allow to access releases as sdsstools.releases. This is important to
-# define it as an extension in Sphinx.
-sys.modules['sdsstools.releases'] = releases